rename js/renewals to js/advantage
