+++ static/js/src/advantage/set-modal-info.js
